requirements.txt
# ----------------------
# Core Python packages
# ----------------------
numpy
pandas
tqdm

# ----------------------
# Image processing
# ----------------------
opencv-python
scikit-image

# ----------------------
# Machine learning & DL
# ----------------------
tensorflow
scikit-learn
xgboost

# ----------------------
# Visualization
# ----------------------
matplotlib
seaborn

# ----------------------
# API (choose Flask or FastAPI)
# ----------------------
flask
python-multipart
# fastapi
# uvicorn

# ----------------------
# BI Dashboard
# ----------------------
streamlit

# ----------------------
# Cloud Integration
# ----------------------
boto3                    # AWS S3
google-cloud-storage     # GCP GCS
azure-storage-blob       # Azure Blob

# ----------------------
# Misc & Monitoring
# ----------------------
requests
